Finish addDataInitial File

+++ src/main/java/fr/picom/picomspring/model/Stop.java
@@ -11,9 +11,9 @@ public class Stop {
 
     private String name;
 
-    private Float latitude;
+    private Double latitude;
 
-    private Float longitude;
+    private Double longitude;
 
     @ManyToOne
     private Area area;
@@ -39,19 +39,19 @@ public class Stop {
         this.name = name;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
